Allow scripts to specify libraries to link against.
Author must add a link-declaration (anywhere in the code), and
the compiled script will be linked against the specified library.

<link-declaration> = //link <library-name>
<library-name> = lib<name>

where name is, e.g. "curl".  So, a full example would look thus:
//link: libcurl

Not sure if this is how I'll keep it ultimately, but it was the
first idea that popped into my head and it seems to be pretty easy
to remember.. we'll see.