Initial commit; rework of how my dotfiles work
I was inspired by a few ideas I've seen around the interwebs
and decided that doing the whole --work-tree=$HOME thing was
a band-wagon I wanted to jump onto.

The main divergent point of my stuff is that I often have a need
to "layer" config settings depending on my work setting.  Therefore,
in addition to using the --work-tree idea, this stuff also eases
creation of additional configuration repositories and the main
rcs (provided by *this* repo) does its best to source in more granular
settings when they are present.